split launcher and logic
