+++ tf2_bot_detector/Launcher/DLLMain.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "tf2_bot_detector_export.h"
+
+namespace tf2_bot_detector
+{
+	TF2_BOT_DETECTOR_EXPORT int RunProgram(int argc, const char** argv);
+}
